aula_1/live/pyproject.toml
[project]
name = "live"
version = "0.1.0"
description = ""
authors = [
    {name = "EonCyber",email = "eonciber@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "keyflow (>=1.1.1,<2.0.0)",
    "keyboard (>=0.13.5,<0.14.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "langchain-core (>=0.3.64,<0.4.0)",
    "langchain-openai (>=0.3.21,<0.4.0)"
]

[tool.poetry]
packages = [{include = "live", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
